// src/components/types.ts
export type Tdata = {
  avatar_url?: string;
  bio?: string | null;
  blog?: string;
  login?: string;
  email?: string | null;
  events_url?: string;
  followers?: number;
  followers_url?: String;
  following?: number;
  following_url?: string;
  gists_url?: string;
  gravatar_id?: string;
  hireable?: number | string | null;
  id?: number;
  location?: string;
  name?: string;
  node_id?: string;
  organizations_url?: string;
  public_gists?: number;
  company?: string;
  twitter_username?: string | null;
  public_repos?: number;
  received_events_url?: string;
  repos_url?: string;
  site_admin?: boolean;
  starred_url?: string;
  subscriptions_url?: string;
  created_at?: string;
  type?: string;
  updated_at?: string;
  url?: string;
};